+++ src/device/tsp.script/Script/SessionBaseExtensions/FirmwareManager-load.cs
@@ -218,8 +218,8 @@ public static partial class FirmwareManager
             if ( tspFile is null )
                 throw new System.IO.FileNotFoundException( "Failed opening TSP Script file", resourceFilePath );
 
-            string trimmedFileSuffix = ".trimmed.tsp";
-            string trimmedFilePath = resourceFilePath.Replace( ".tsp", trimmedFileSuffix );
+            string trimmedFileSuffix = $".trimmed{FirmwareScriptBase.ScriptFileExtension}";
+            string trimmedFilePath = resourceFilePath.Replace( FirmwareScriptBase.ScriptFileExtension, trimmedFileSuffix );
             if ( string.Equals( trimmedFilePath, resourceFilePath ) )
                 trimmedFilePath = resourceFilePath + trimmedFileSuffix;